requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# AWS Bedrock + services
langchain-aws>=0.1.0
boto3>=1.34.0

# Vector DB for RAG (local, free)
chromadb>=0.4.0

# UI & config
streamlit>=1.30.0
python-dotenv>=1.0.0

# Production Redshift (optional)
# psycopg2-binary>=2.9.0
